==> code/03_Feature_Engineering.ipynb <==
# Library imports, including MinMaxScaler from SKLearn

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.preprocessing import MinMaxScaler

# Max columns and rows for easier viewing of certain operations below

# pd.set_option('display.max_columns', None)
# pd.set_option('display.max_rows', None)

# Read in cleaned CSVs into two different dataframes

df_train_cleaned = pd.read_csv('../data/train_cleaned.csv')
df_test_cleaned = pd.read_csv('../data/test_cleaned.csv')

# View header of train_cleaned dataframe to get an idea of the data and columns

df_train_cleaned.head()

After importing the libraries and reading in the data, I ran descriptive statistics again (below) for all numerical columns. Considering I was going to log several features, I needed to first scale any categories to remove 0 values. Using the .describe() below I was able to easily identify which features I would need to scale. I chose to use the MinMax Scaler instead of the StandardScaler since the StandardScaler assumes the data follows a normal distribution which not all my data did.

# Take a look at the descriptive statistics for the training dataset to understand which values
# I would need to scale in order to perform meaningful transformations or logs

df_train_cleaned.describe()

In the following cell I performed all of my feature engineering operations. I initially started with performing a linear shift for certain columns so that I could either apply a log to them or use them in an interaction term. The reason I just performed a linear shift on this data is because they were all values that were less than ten and were hard counts of each feature. I then used the MinMaxScaler on features that I wanted to log or interact with that were also much more expansive in their ranges or may have outliers. I then created interaction term categories by multiplying different categories that I thought might work well together. Lastly, I logged features with a large spread or possible outliers to examine when testing my model in the next notebook.

# Feature engineering function
# https://www.analyticsvidhya.com/blog/2020/07/types-of-feature-transformation-and-scaling/
# https://stackoverflow.com/questions/45554008/error-in-python-script-expected-2d-array-got-1d-array-instead

def feat_eng(df):
    # Features for linear shift to get rid of 0s
    df['full_bath_shift'] = df['full_bath'] + 1
    df['half_bath_shift'] = df['half_bath'] + 1
    df['bedrooms_gr_shift'] = df['bedrooms_gr'] + 1
    df['fireplaces_shift'] = df['fireplaces'] + 1
    df['garage_car_size_shift'] = df['garage_car_size'] + 1
        
    # Features to MinMaxScale (1, 2)
    scaler = MinMaxScaler(feature_range=(1, 2))
    df['lot_area_scaled_mm'] = scaler.fit_transform(df[['lot_area']])
    df['bsmt_fin_1_sf_scaled_mm'] = scaler.fit_transform(df[['bsmt_fin_1_sf']])
    df['bsmt_fin_2_sf_scaled_mm'] = scaler.fit_transform(df[['bsmt_fin_2_sf']])
    df['bsmt_sf_scaled_mm'] = scaler.fit_transform(df[['bsmt_sf']])
    df['gr_liv_area_scaled_mm'] = scaler.fit_transform(df[['gr_liv_area']])
    df['wood_deck_sf_scaled_mm'] = scaler.fit_transform(df[['wood_deck_sf']])
    
    # Feature engineering / multiplying/adding columns
    df['tot_rooms_gr_gr_liv_area'] = df['tot_rooms_gr'] * df['gr_liv_area']
    df['tot_rooms_gr_bedroom_gr'] = df['tot_rooms_gr'] * (df['bedrooms_gr_shift'])
    df['full_bath_gr_liv_area'] = (df['full_bath_shift']) * df['gr_liv_area']
    df['gar_car_size_overall_qual'] = df['garage_car_size_shift'] * df['overall_qual']
    df['gar_car_size_overall_qual'] = df['garage_car_size_shift'] * df['overall_qual']
    df['yr_built_overall_qual'] = df['yr_built'] * df['overall_qual']
    df['bsmt_sf_gr_liv_area'] = df['bsmt_sf'] + df['gr_liv_area']
    df['functional_mapped_overall_qual'] = df['functional_mapped'] * df['overall_qual']
    
    # Features to convert to logs
    df['full_bath_gr_liv_area_log'] = df['full_bath_gr_liv_area'].map(np.log)
    df['tot_rooms_gr_bedroom_gr_log'] = df['tot_rooms_gr_bedroom_gr'].map(np.log)
    df['lot_area_scaled_mm_log'] = df['lot_area_scaled_mm'].map(np.log)
    df['gr_liv_area_scaled_mm_log'] = df['gr_liv_area_scaled_mm'].map(np.log)
    df['bsmt_fin_1_sf_scaled_mm_log'] = df['bsmt_fin_1_sf_scaled_mm'].map(np.log)
    df['bsmt_fin_2_sf_scaled_mm_log'] = df['bsmt_fin_2_sf_scaled_mm'].map(np.log)
    df['bsmt_sf_scaled_mm_log'] = df['bsmt_sf_scaled_mm'].map(np.log)
    df['bedrooms_gr_shift_log'] = df['bedrooms_gr_shift'].map(np.log)
    df['tot_rooms_gr_log'] = df['tot_rooms_gr'].map(np.log)
    df['tot_rooms_gr_gr_liv_area_log'] = df['tot_rooms_gr_gr_liv_area'].map(np.log)
    df['lot_area_log'] = df['lot_area'].map(np.log)
    df['gr_liv_area_log'] = df['gr_liv_area'].map(np.log)
    df['wood_deck_sf_scaled_mm_log'] = df['wood_deck_sf_scaled_mm'].map(np.log)
    
    return df

# Run my training data through the feat_eng() function and view head

df_train_cleaned_eng = feat_eng(df_train_cleaned)
df_train_cleaned_eng.head()

# Run my test data through the feat_eng() function and view head

df_test_cleaned_eng = feat_eng(df_test_cleaned)
df_test_cleaned_eng.head()

After running my training and testing data through the feat_eng() function I saved the output to new CSV files.

# Export engineered data to CSVs for model selection and tuning

df_test_cleaned.to_csv('../data/test_cleaned_eng.csv', index=False)
df_train_cleaned.to_csv('../data/train_cleaned_eng.csv', index=False)
