==> code/05_Regularization.ipynb <==
# Import libraries

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.linear_model import LinearRegression, Ridge, RidgeCV, Lasso, LassoCV
from sklearn import metrics
from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.preprocessing import StandardScaler

For the Regularization models, I first loaded the feature selected dataframe and my y target dataframe from my Modelling notebook. I then performed the train/test/split function and fit_transformed the X_train split and transformed the X_validation split.

# Load necessary dataframes

X_train_df = pd.read_csv('../data/X_train_df.csv')
y = pd.read_csv('../data/y.csv')

# Train/test/split, scale, and fit_transform/transform the data

X_train, X_validation, y_train, y_validation = train_test_split(X_train_df, y, random_state=42)
sc = StandardScaler()            

scalar_columns = list(X_train.select_dtypes(exclude='object'))

Z_train = sc.fit_transform(X_train[scalar_columns])
Z_validation = sc.transform(X_validation[scalar_columns]) 

I created my alphas, instantiated and cross-validated the RidgeCV model, and fit my RidgeCV model. I checked the RidgeCV alpha and printed the test scores. I then repeated the process for my LassoCV model and printed the RidgeCV and LassoCV coefficients.

# Create alphas, instantiate and cross-validate, and fit

r_alphas = np.logspace(0, 5, 100)   # Get 100 values from 10^0 to 10^5
ridge_cv = RidgeCV(alphas=r_alphas, scoring='r2', cv=5)   # 5-fold CV
ridge_cv.fit(Z_train, y_train);

# Find the optimal alpha value

ridge_cv.alpha_
# ridge_cv.best_score_ 

# Print RidgeCV scores

print('Training RidgeCV score: ', ridge_cv.score(Z_train, y_train))
print('Test RidgeCV score: ', ridge_cv.score(Z_validation, y_validation))

# Create alphas, instantiate and cross-validate, and fit

l_alphas = np.logspace(-3, 0, 100)   # 10^-3 = .001 up to 10^0 = 1
lasso_cv = LassoCV(alphas=l_alphas, cv=5, max_iter=50000)
lasso_cv.fit(Z_train, y_train.values.flatten());

# Find the optimal alpha value

lasso_cv.alpha_

# Print LassoCV scores

print('Training LassoCV score: ', lasso_cv.score(Z_train, y_train))
print('Test LassoCV score: ', lasso_cv.score(Z_validation, y_validation))

ridge_cv.coef_

lasso_cv.coef_
